refactor(github): move GithubDeployment from github_graphql to github
