+++ backend/plugins/github/models/migrationscripts/register.go
@@ -43,5 +43,6 @@ func All() []plugin.MigrationScript {
 		new(addGithubIssueAssignee),
 		new(addFullName),
 		new(addRawParamTableForScope),
+		new(addDeploymentTable),
 	}
 }
